requirements.txt
fastapi
motor
pydantic
python-dotenv   
uvicorn 
pytest
